+++ CHANGELOG.md
@@ -14,6 +14,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -   New "Report Issue" command (#93)
 -   New "Show Output" command
 -   Extend OutputChannel to be able to buffer output internally for error reporting (up to 1000 lines)
+-   Add button to report server crashes
+-   Abstract out Max Restart Count into a setting `psalm.maxRestartCount`
 
 ## [2.2.3] - 2021-09-17